In ROOT source code we banish the trailing blanks at the end of the lines.
Therefore most of the editors of people working as ROOT developers
are setup to remove automatically trailing blanks each time a file
is saved.

This prevent to use the markdown syntax for the line break which is two
blanks at the end of a line. We should prefer the traditional html <br> instead.