Add a script to generate tables of build options.
New script in .github/scripts/ that generates a table of build options,
and writes them to _include/.
It takes one or more branch names, and generates one or more tables.

When not run in web/, one has to pass the location of _include/ for it
to work.

The script recognises both ROOT_BUILD_OPTION( ) and cmake-style option( )