Link checker: ignore 'root.cern/[0-9]+', not 'ph-root-2.cern.ch/.*'
We would rather not have links to ph-root-2 around the website, so let's
not ignore them. Instead, now that root.cern is the configured site
name, we must ignore URLs with the form 'root.cern/[0-9]+' because they
appear in the html pages' metadata, but are always broken since we run
the checker on a local build of the website and not on a deployment.