Added directions for updating/using the `latest-stable` branch
The `latest-stable` branch is a convenient way for users that build from source to get
the latest revision that is considered stable by the ROOT team.

The name of this branch does not change, meaning that a `git clone -b latest-stable`
will fetch the the most recent revision without manually specifying a tag name.
Also, this branch is meant to be updated regularly as part of the release
procedure. As such, users can upgrade to the latest published version by
issuing a simple `git pull`.

TL;DR
We attain this by manually committing (via `git commit-tree`) the tree of the
latest stable version to the `latest-stable` branch. Basically, we create a linear
history in which each commit points to the tree of the latest tagged release.

On the client-side, for a `git pull` this is always seen as a fast-forward, and
hence the update is trivial, not requiring a merge.