Don't mention STL backports and don't list minimum C++ standards twice
The STL backports should not be mentioned, because this information will
be quickly outdated. The link is already dead, the `std::string_view`
backport is already gone, and the remaining `std::span` backport will
also be gone with the next increment of the required minimual C++
standard.

Also, organize the information a bit better, avoiding redundant
explanation of the C++ standard caveat and minimum supported C++
standards.